requirements.txt
Flask
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
Flask-Session
python-dotenv
requests
python-dateutil
mysql-connector-python>=9.0.0
